[floats] use floats in span.Metrics

+++ model/span.go
@@ -23,9 +23,9 @@ type Span struct {
 	Meta map[string]string `json:"meta"` // arbitrary tags/metadata
 
 	// TODO float64
-	Metrics  map[string]int64 `json:"metrics"`   // arbitrary metrics
-	ParentID uint64           `json:"parent_id"` // span ID of the span in which this one was created
-	Type     string           `json:"type"`      // protocol associated with the span
+	Metrics  map[string]float64 `json:"metrics"`   // arbitrary metrics
+	ParentID uint64             `json:"parent_id"` // span ID of the span in which this one was created
+	Type     string             `json:"type"`      // protocol associated with the span
 }
 
 // String formats a Span struct to be displayed as a string
